backend/requirements.txt
fastapi[standard]
uvicorn
beautifulsoup4
zendriver
google-generativeai
pillow
